requirements.txt
flask
openai
pillow
python-dotenv
gunicorn
